# Optimize the infer program and run for ocr attention model.