… OCR/paddle/outputs/res_infer_cpu_perf.txt
… OCR/paddle/outputs/res_infer_gpu_perf.txt