- optimize the infer program: 23ms -> 20ms (GPU infer time)
- set `use_program_cache=True` in Executor.run: 20ms -> 12ms (GPU infer time)